+++ observer/matrix_to_grid_test.go
@@ -11,16 +11,24 @@ import (
 
 type matObs struct {
 	values []float64
+	Cols   int
+	Rows   int
 }
 
 func (o *matObs) Initialize(w *ecs.World) {
-	o.values = make([]float64, 20*30)
+	if o.Cols == 0 {
+		o.Cols = 30
+	}
+	if o.Rows == 0 {
+		o.Rows = 20
+	}
+	o.values = make([]float64, o.Cols*o.Rows)
 }
 
 func (o *matObs) Update(w *ecs.World) {}
 
 func (o *matObs) Dims() (int, int) {
-	return 30, 20
+	return o.Cols, o.Rows
 }
 
 func (o *matObs) Values(w *ecs.World) []float64 {
@@ -31,11 +39,11 @@ func TestMatrixToGrid(t *testing.T) {
 	m := model.New()
 
 	var mat observer.Matrix = &matObs{}
-	var grid observer.Grid = &observer.MatrixToGrid{
-		Observer: mat,
-		Origin:   [...]float64{1, 2},
-		CellSize: [...]float64{5, 10},
-	}
+	var grid observer.Grid = observer.MatrixToGrid(
+		mat,
+		&[...]float64{1, 2},
+		&[...]float64{5, 10},
+	)
 
 	grid.Initialize(&m.World)
 	grid.Update(&m.World)
@@ -53,10 +61,11 @@ func TestMatrixToGrid(t *testing.T) {
 	assert.Equal(t, 6.0, grid.X(1))
 	assert.Equal(t, 12.0, grid.Y(1))
 
-	grid = &observer.MatrixToGrid{
-		Observer: mat,
-	}
+	grid = observer.MatrixToGrid(mat, nil, nil)
 	grid.Initialize(&m.World)
 	assert.Equal(t, 1.0, grid.X(1))
 	assert.Equal(t, 1.0, grid.Y(1))
+
+	data := grid.Values(&m.World)
+	assert.Equal(t, 20*30, len(data))
 }